move entities to be exposed
