+++ internal/converter/resource.go
@@ -22,7 +22,7 @@ import (
 	"strings"
 
 	"github.com/whiteblock/definition/config/defaults"
-	"github.com/whiteblock/definition/internal/entity"
+	"github.com/whiteblock/definition/pkg/entity"
 	"github.com/whiteblock/definition/schema"
 
 	util "github.com/whiteblock/utility/utils"
