+++ internal/distribute/biome.go
@@ -20,8 +20,8 @@ package distribute
 
 import (
 	"github.com/whiteblock/definition/config"
-	"github.com/whiteblock/definition/internal/entity"
 	"github.com/whiteblock/definition/internal/parser"
+	"github.com/whiteblock/definition/pkg/entity"
 	"github.com/whiteblock/definition/schema"
 
 	"github.com/sirupsen/logrus"
