+++ internal/distribute/biome_test.go
@@ -22,10 +22,10 @@ import (
 	"testing"
 
 	"github.com/whiteblock/definition/config"
-	"github.com/whiteblock/definition/internal/entity"
 	mockDist "github.com/whiteblock/definition/internal/mocks/distribute"
 	mockEntity "github.com/whiteblock/definition/internal/mocks/entity"
 	mockParser "github.com/whiteblock/definition/internal/mocks/parser"
+	"github.com/whiteblock/definition/pkg/entity"
 	"github.com/whiteblock/definition/schema"
 
 	"github.com/sirupsen/logrus"
