+++ internal/distribute/distribute.go
@@ -20,7 +20,7 @@ package distribute
 
 import (
 	"github.com/whiteblock/definition/config"
-	"github.com/whiteblock/definition/internal/entity"
+	"github.com/whiteblock/definition/pkg/entity"
 	"github.com/whiteblock/definition/schema"
 
 	"github.com/sirupsen/logrus"
