+++ internal/distribute/state.go
@@ -22,10 +22,10 @@ import (
 	"errors"
 	"fmt"
 
-	"github.com/whiteblock/definition/internal/entity"
 	"github.com/whiteblock/definition/internal/merger"
 	"github.com/whiteblock/definition/internal/namer"
 	"github.com/whiteblock/definition/internal/parser"
+	"github.com/whiteblock/definition/pkg/entity"
 	"github.com/whiteblock/definition/schema"
 )
