+++ internal/parser/resources.go
@@ -24,9 +24,9 @@ import (
 	"strings"
 
 	"github.com/whiteblock/definition/internal/converter"
-	"github.com/whiteblock/definition/internal/entity"
 	"github.com/whiteblock/definition/internal/namer"
 	"github.com/whiteblock/definition/internal/search"
+	"github.com/whiteblock/definition/pkg/entity"
 	"github.com/whiteblock/definition/schema"
 )
