+++ internal/parser/service.go
@@ -25,8 +25,8 @@ import (
 	"github.com/whiteblock/definition/command"
 	"github.com/whiteblock/definition/config/defaults"
 	"github.com/whiteblock/definition/internal/converter"
-	"github.com/whiteblock/definition/internal/entity"
 	"github.com/whiteblock/definition/internal/namer"
+	"github.com/whiteblock/definition/pkg/entity"
 	"github.com/whiteblock/definition/schema"
 )
