+++ internal/parser/sidecar.go
@@ -23,8 +23,8 @@ import (
 
 	"github.com/whiteblock/definition/config/defaults"
 	"github.com/whiteblock/definition/internal/converter"
-	"github.com/whiteblock/definition/internal/entity"
 	"github.com/whiteblock/definition/internal/namer"
+	"github.com/whiteblock/definition/pkg/entity"
 	"github.com/whiteblock/definition/schema"
 
 	"github.com/jinzhu/copier"
