+++ internal/process/process.go
@@ -21,7 +21,7 @@ package process
 import (
 	"fmt"
 
-	"github.com/whiteblock/definition/internal/entity"
+	"github.com/whiteblock/definition/pkg/entity"
 	"github.com/whiteblock/definition/schema"
 )
