+++ internal/process/process_test.go
@@ -21,8 +21,8 @@ package process
 import (
 	"testing"
 
-	"github.com/whiteblock/definition/internal/entity"
 	"github.com/whiteblock/definition/internal/mocks/process"
+	"github.com/whiteblock/definition/pkg/entity"
 	"github.com/whiteblock/definition/schema"
 
 	"github.com/stretchr/testify/assert"
