+++ pkg/entity/image_store.go
