+++ pkg/entity/network.go
