+++ pkg/entity/network_state.go
