+++ pkg/entity/network_state_test.go
