+++ pkg/entity/network_test.go
