+++ pkg/entity/segment.go
