+++ pkg/entity/service.go
