+++ pkg/entity/state.go
