+++ pkg/entity/system_diff.go
